requirements.txt
python-dotenv
geopy
requests
sqlalchemy
psycopg2-binary